agent-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
